Implement auto increment ins DB-specific way

+++ src/main/kotlin/io/github/guttenbase/meta/InternalTableMetaData.kt
@@ -10,8 +10,8 @@ package io.github.guttenbase.meta
 internal interface InternalTableMetaData : TableMetaData {
   override var filteredRowCount: Int
   override var totalRowCount: Int
-  override var maxId: Int
-  override var minId: Int
+  override var maxId: Long
+  override var minId: Long
 
   fun addColumn(column: ColumnMetaData)
   fun removeColumn(columnMetaData: ColumnMetaData)
